+++ service_catalog/models/request_state.py
@@ -3,7 +3,7 @@ from django.db.models import IntegerChoices
 
 class RequestState(IntegerChoices):
     SUBMITTED = 1, "SUBMITTED"
-    NEED_INFO = 2, "NEED_INFO"
+    ON_HOLD = 2, "ON_HOLD"
     REJECTED = 3, "REJECTED"
     CANCELED = 4, "CANCELED"
     ACCEPTED = 5, "ACCEPTED"