+++ tests/test_service_catalog/test_api/test_request/test_request_patch.py
@@ -25,7 +25,7 @@ class TestApiRequestPatch(BaseTestRequestAPI):
         )
         self.patch_data = {
             'operation': self.update_operation_test.id,
-            'state': RequestState.NEED_INFO,
+            'state': RequestState.ON_HOLD,
         }
         self.kwargs = {
             'pk': self.test_request_standard_user_1.id
@@ -35,7 +35,7 @@ class TestApiRequestPatch(BaseTestRequestAPI):
             'id': self.test_request_standard_user_1.id,
             'fill_in_survey': self.test_request_standard_user_1.fill_in_survey,
             'tower_job_id': self.test_request_standard_user_1.tower_job_id,
-            'state': RequestState.NEED_INFO,
+            'state': RequestState.ON_HOLD,
             'operation': self.update_operation_test.id,
             'user': UserSerializerNested(self.test_request_standard_user_1.user).data
         }