+++ tests/test_service_catalog/test_api/test_request/test_request_state_machine/test_cancel.py
@@ -3,7 +3,7 @@ from django.urls import reverse
 from service_catalog.models import RequestState, InstanceState
 from tests.test_service_catalog.base_test_request import BaseTestRequestAPI
 
-AUTHORIZED_STATES = [RequestState.SUBMITTED, RequestState.NEED_INFO, RequestState.REJECTED, RequestState.ACCEPTED]
+AUTHORIZED_STATES = [RequestState.SUBMITTED, RequestState.ON_HOLD, RequestState.REJECTED, RequestState.ACCEPTED]
 
 
 class TestApiRequestCancel(BaseTestRequestAPI):